requirements.txt
python-docx==1.1.0
openpyxl==3.1.2
reportlab==4.0.7
jinja2==3.1.3
pandas==2.1.4
python-dateutil==2.8.2
PyPDF2==3.0.1
fpdf2==2.7.6
flask==3.0.0
werkzeug==3.0.1

